fix: check file size on drag and drop + refactor
