+++ src/Template/Layout/js/app/pages/import/index.js
@@ -3,10 +3,7 @@
  *  Template/Import/index.twig
  *
  * <modules-view> component used for ModulesPage -> View
- *
  */
-import { warning } from 'app/components/dialog/dialog';
-import { t } from 'ttag';
 
 export default {
     props: {
@@ -47,17 +44,11 @@ export default {
 
     methods: {
         onFileChanged(e) {
-            const fileSize = Math.round(e.target.files[0].size);
-            const maxFileSize = BEDITA.maxFileSize;
-            if (fileSize >= maxFileSize) {
-                this.fileName = '';
-                const fileSizeMb = Math.round(fileSize / 1024 / 1024);
-                const maxFileSizeMb = Math.round(BEDITA.maxFileSize / 1024 / 1024);
-                const message = t`File too big (${fileSizeMb} MB), please select a file less than max file size (${maxFileSizeMb} MB)`;
-                warning(message);
-
+            this.fileName = '';
+            if (this.$helpers.checkMaxFileSize(e.target.files[0]) === false) {
                 return;
             }
+
             this.fileName = e.target.files[0].name;
         },
